Move array stream parsing funcs to array modules
